daemon - improve perf by rewriting insert-on-duplicate from db to daemon
this attempts to fix a long standing performance issue in the retrieval
of results for an input. For example: result_php73 is currently 5.5 GiB
with a 1975 MiB inputVersion UNIQUE idx. Since the results are clustered
by input, we don't need Version in that idx, it's only there for the
unique. This change removes the need for that idx, meaning we can have a
simple idx on input (which needs only 700 MiB).

While I still don't understand the absolute sizes for pg's indexes
(result_php73 currently has 92M rows for 4.5M inputs) - this change
trims the idx by 2/3rd without any negative effect